requirements.txt
flask
flask-sqlalchemy
flask-migrate
psycopg2-binary
flask-restx
python-pptx
apscheduler